Create dictionary.py
Quiz: Verse Dictionary
In the code editor below, you'll find a dictionary containing the unique words of verse stored as keys and the number of times they appear in verse stored as values. Use this dictionary to answer the following questions. Submit these answers in the quiz below the code editor.

Try to answer these using code, rather than inspecting the dictionary manually!

How many unique words are in verse_dict?
Is the key "breathe" in verse_dict?
What is the first element in the list created when verse_dict is sorted by keys?
Hint: Use the appropriate dictionary method to get a list of its keys, and then sort that list. Use this list of keys to answer the next two questions as well.
Which key (word) has the highest value in verse_dict?