Create dictionaries_loops.py
Building Dictionaries
two important concepts: 1) counting with for loops and 2) the dictionary get method. These two can actually be combined to create a useful counter dictionary, something you will likely come across again. For example, we can create a dictionary, word_counter, that keeps track of the total count of each word in a string.

Two important ways : 
Method 1: Using a for loop to create a set of counters (Iterate through each element in the list. If an element is already included in the dictionary, add 1 to its value. If not, add the element to the dictionary and set its value to 1.)
Method 2: Using the get method (Iterate through each element, get() its value in the dictionary, and add 1.)